// WADNR.Web/src/app/shared/generated/model/tag-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TagUpsertRequest { 
    TagName?: string | null;
    TagDescription?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TagUpsertRequestForm { 
    TagName?: FormControl<string>;
    TagDescription?: FormControl<string>;
}

export class TagUpsertRequestFormControls { 
    public static TagName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TagDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}